pyproject.toml
[project]
name = "blogger-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langgraph-cli[inmem]>=0.4.12",
    "uvicorn>=0.40.0",
    "watchdog>=6.0.0",
]
